requirements.txt
requests
flask
# json
beautifulsoup4
python-marmiton
pymongo